CMS-719 Sync-key are not taken into account during import
The underlying problem was that the default count for the query fetching the existing content sync-keys was cutting off the set at 200 while the actual number of content could be anything.

Changed the default count to be the threshold where an extra query to fetch the total number of hits are used.